tresure_hunt/requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
Flask-Limiter==3.7.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
Pillow==10.4.0
